requirements.txt
aiosmtplib==1.1.7
anyio==3.6.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.11
asttokens==2.0.5
attrs==22.1.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
black==22.10.0
bleach==4.1.0
blinker==1.5
Brlapi==0.7.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==38.0.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dell-recovery==0.0.0
DellLinuxAssistant==0.0.0
dill==0.3.5.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
duplicity==0.8.12.0
email-validator==1.3.0
entrypoints==0.3
executing==0.8.3
fastapi==0.85.0
fastapi-mail==1.2.0
fasteners==0.14.1
filelock==3.8.0
findspark==2.0.1
Flask==2.0.3
flask-unsign==1.2.0
future==0.18.2
gpg===1.13.1-unknown
greenlet==0.4.15
h11==0.14.0
hashID==3.1.4
httplib2==0.14.0
idna==3.4
impacket==0.9.24
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.4
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jwt==1.3.1
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
ldap3==2.9.1
ldapdomaindump==0.9.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mccabe==0.7.0
metakernel==0.29.0
mistune==0.8.4
monotonic==1.5
msgpack==0.6.2
mypy-extensions==0.4.3
nbclassic==0.3.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.20.3
oauthlib==3.1.0
packaging==21.3
pandas==1.4.2
pandas-datareader==0.10.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.10.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
pony==0.7.16
progressbar==2.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.10.2
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==2.5.0
pylint==2.15.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.1.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.21.0
python-magic==0.4.16
python-multipart==0.0.5
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.28.1
requests-unixsocket==0.2.0
scipy==1.6.3
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smtp-user-enum==0.5.0
sniffio==1.3.0
sos==4.3
soupsieve==2.3.1
SpeechRecognition==3.8.1
spylon==0.3.0
spylon-kernel==0.4.1
sqlmap==1.6.8
ssh-import-id==5.10
stack-data==0.2.0
starlette==0.20.4
systemd-python==234
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tornado==6.1
traitlets==5.1.1
typing-extensions==4.4.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
uvicorn==0.19.0
sphinx-rtd-theme
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.4
Werkzeug==2.0.3
wfuzz==2.4.5
wrapt==1.14.1
xkit==0.0.0
zipp==3.7.0